♻️Refactor: MenuBar 리팩토링
