+++ src/Components/common/Header.jsx
@@ -1,4 +1,4 @@
-import MenuBar from "../MenuBar";
+import MenuBar from "../MenuBar/MenuBar";
 
 export default function Header() {
   return (
